Break out input and output types into separate files
