+++ inst/include/general_demographic_projection.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "types.hpp"
+#include "state_types.hpp"
 
 namespace leapfrog {
