+++ inst/include/hiv_demographic_projection.hpp
@@ -2,6 +2,7 @@
 
 #include <iostream>
 #include "types.hpp"
+#include "state_types.hpp"
 
 namespace leapfrog {
