+++ inst/include/types.hpp
@@ -75,255 +75,10 @@ constexpr int hPS = StateSpace<ModelVariant>().children.hPS;
 template<typename ModelVariant>
 constexpr int hBF = StateSpace<ModelVariant>().children.hBF;
 
-template<typename real_type>
-struct Options {
-  int hts_per_year;
-  double dt;
-  const int p_idx_fertility_first;
-  const int p_fertility_age_groups;
-  const int p_idx_hiv_first_adult;
-  const int adult_incidence_first_age_group;
-  const int pAG_INCIDPOP;
-  const int ts_art_start;
-  const int hAG_15plus;
-  const int hIDX_15PLUS;
-
-  Options(int hts_per_year,
-          int ts_art_start,
-          int hAG_15plus) :
-      hts_per_year(hts_per_year),
-      dt(1.0 / hts_per_year),
-      p_idx_fertility_first(15),
-      p_fertility_age_groups(35),
-      p_idx_hiv_first_adult(15),
-      adult_incidence_first_age_group(15),
-      pAG_INCIDPOP(35),
-      ts_art_start(ts_art_start),
-      hAG_15plus(hAG_15plus),
-      hIDX_15PLUS(0) {}
-};
-
-template<typename real_type>
-struct Demography {
-  TensorMap2<real_type> base_pop;
-  TensorMap3<real_type> survival_probability;
-  TensorMap3<real_type> net_migration;
-  TensorMap2<real_type> age_specific_fertility_rate;
-  TensorMap2<real_type> births_sex_prop;
-};
-
-template<typename real_type>
-struct Incidence {
-  TensorMap1<real_type> total_rate;
-  TensorMap3<real_type> relative_risk_age;
-  TensorMap1<real_type> relative_risk_sex;
-};
-
-template<typename real_type>
-struct NaturalHistory {
-  TensorMap3<real_type> cd4_mortality;
-  TensorMap3<real_type> cd4_progression;
-  TensorMap3<real_type> cd4_initial_distribution;
-  int scale_cd4_mortality;
-};
-
-template<typename real_type>
-struct Art {
-  Tensor1<int> idx_hm_elig;
-  TensorMap4<real_type> mortality;
-  TensorMap2<real_type> mortaility_time_rate_ratio;
-  Tensor1<real_type> h_art_stage_dur;
-  TensorMap1<real_type> dropout;
-  TensorMap2<real_type> adults_on_art;
-  TensorMap2<int> adults_on_art_is_percent;
-  real_type initiation_mortality_weight;
-};
-
-template<typename real_type>
-struct BaseModelParameters {
-  Options<real_type> options;
-  Demography<real_type> demography;
-  Incidence<real_type> incidence;
-  NaturalHistory<real_type> natural_history;
-  Art<real_type> art;
-};
-
-template<typename real_type>
-struct Children {
-  TensorMap1<real_type> hc_nosocomial;
-  TensorMap1<real_type> hc1_cd4_dist;
-  TensorMap2<real_type> hc_cd4_transition;
-  TensorMap3<real_type> hc1_cd4_mort;
-  TensorMap3<real_type> hc2_cd4_mort;
-  TensorMap1<real_type> hc1_cd4_prog;
-  TensorMap1<real_type> hc2_cd4_prog;
-  real_type ctx_effect;
-  TensorMap1<real_type> ctx_val;
-  TensorMap1<real_type> hc_art_elig_age;
-  Tensor2<real_type> hc_art_elig_cd4;
-  TensorMap3<real_type> hc_art_mort_rr;
-  TensorMap3<real_type> hc1_art_mort;
-  TensorMap3<real_type> hc2_art_mort;
-  TensorMap1<int> hc_art_isperc;
-  TensorMap1<real_type> hc_art_val;
-  TensorMap2<real_type> hc_art_init_dist;
-};
-
-template<typename ModelVariant, typename real_type>
-struct ChildModelParameters {
-};
-
-template<typename real_type>
-struct ChildModelParameters<ChildModel, real_type> {
-  Children<real_type> children;
-};
-
-template<typename ModelVariant, typename real_type>
-struct Parameters {
-  BaseModelParameters<real_type> base;
-  ChildModelParameters<ModelVariant, real_type> children;
-};
-
 namespace {
 using Eigen::Sizes;
 using Eigen::TensorFixedSize;
 }
-
-template<typename ModelVariant, typename real_type>
-struct BaseModelState {
-  TensorFixedSize <real_type, Sizes<pAG<ModelVariant>, NS<ModelVariant>>> p_total_pop;
-  TensorFixedSize <real_type, Sizes<pAG<ModelVariant>, NS<ModelVariant>>> p_total_pop_natural_deaths;
-  TensorFixedSize <real_type, Sizes<pAG<ModelVariant>, NS<ModelVariant>>> p_hiv_pop;
-  TensorFixedSize <real_type, Sizes<pAG<ModelVariant>, NS<ModelVariant>>> p_hiv_pop_natural_deaths;
-  TensorFixedSize <real_type, Sizes<hDS<ModelVariant>, hAG<ModelVariant>, NS<ModelVariant>>> h_hiv_adult;
-  TensorFixedSize <real_type, Sizes<hTS<ModelVariant>, hDS<ModelVariant>, hAG<ModelVariant>, NS<ModelVariant>>>
-      h_art_adult;
-  real_type births;
-  TensorFixedSize <real_type, Sizes<hDS<ModelVariant>, hAG<ModelVariant>, NS<ModelVariant>>> h_hiv_deaths_no_art;
-  TensorFixedSize <real_type, Sizes<pAG<ModelVariant>, NS<ModelVariant>>> p_infections;
-  TensorFixedSize <real_type, Sizes<hTS<ModelVariant>, hDS<ModelVariant>, hAG<ModelVariant>, NS<ModelVariant>>>
-      h_hiv_deaths_art;
-  TensorFixedSize <real_type, Sizes<hDS<ModelVariant>, hAG<ModelVariant>, NS<ModelVariant>>> h_art_initiation;
-  TensorFixedSize <real_type, Sizes<pAG<ModelVariant>, NS<ModelVariant>>> p_hiv_deaths;
-
-  BaseModelState(const Parameters<ModelVariant, real_type> &pars) {
-    constexpr auto ss = StateSpace<ModelVariant>().base;
-    for (int g = 0; g < ss.NS; ++g) {
-      for (int a = 0; a < ss.pAG; ++a) {
-        p_total_pop(a, g) = pars.base.demography.base_pop(a, g);
-      }
-    }
-    p_total_pop_natural_deaths.setZero();
-    p_hiv_pop.setZero();
-    p_hiv_pop_natural_deaths.setZero();
-    h_hiv_adult.setZero();
-    h_art_adult.setZero();
-    births = 0;
-    h_hiv_deaths_no_art.setZero();
-    p_infections.setZero();
-    h_hiv_deaths_art.setZero();
-    h_art_initiation.setZero();
-    p_hiv_deaths.setZero();
-  }
-
-  void set_initial_state(const Parameters<ModelVariant, real_type> &pars) {
-    constexpr auto ss = StateSpace<ModelVariant>().base;
-    for (int g = 0; g < ss.NS; ++g) {
-      for (int a = 0; a < ss.pAG; ++a) {
-        p_total_pop(a, g) = pars.base.demography.base_pop(a, g);
-      }
-    }
-    p_total_pop_natural_deaths.setZero();
-    p_hiv_pop.setZero();
-    p_hiv_pop_natural_deaths.setZero();
-    h_hiv_adult.setZero();
-    h_art_adult.setZero();
-    births = 0;
-    h_hiv_deaths_no_art.setZero();
-    p_infections.setZero();
-    h_hiv_deaths_art.setZero();
-    h_art_initiation.setZero();
-    p_hiv_deaths.setZero();
-  }
-
-  void reset() {
-    p_total_pop.setZero();
-    p_total_pop_natural_deaths.setZero();
-    p_hiv_pop.setZero();
-    p_hiv_pop_natural_deaths.setZero();
-    h_hiv_adult.setZero();
-    h_art_adult.setZero();
-    h_hiv_deaths_no_art.setZero();
-    p_infections.setZero();
-    h_hiv_deaths_art.setZero();
-    h_art_initiation.setZero();
-    p_hiv_deaths.setZero();
-    births = 0;
-  }
-};
-
-template<typename ModelVariant, typename real_type>
-struct ChildModelState {
-  ChildModelState(const Parameters<ModelVariant, real_type> &pars) {}
-
-  void set_initial_state(const Parameters<ModelVariant, real_type> &pars) {}
-
-  void reset() {}
-};
-
-template<typename real_type>
-struct ChildModelState<ChildModel, real_type> {
-  TensorFixedSize <real_type, Sizes<hc1DS<ChildModel>, hcTT<ChildModel>, hc1AG<ChildModel>, NS<ChildModel>>> hc1_hiv_pop;
-  TensorFixedSize <real_type, Sizes<hc2DS<ChildModel>, hcTT<ChildModel>, hc2AG<ChildModel>, NS<ChildModel>>> hc2_hiv_pop;
-  TensorFixedSize <real_type, Sizes<hTS<ChildModel>, hc1DS<ChildModel>, hc1AG<ChildModel>, NS<ChildModel>>> hc1_art_pop;
-  TensorFixedSize <real_type, Sizes<hTS<ChildModel>, hc2DS<ChildModel>, hc2AG<ChildModel>, NS<ChildModel>>> hc2_art_pop;
-  TensorFixedSize <real_type, Sizes<hc1DS<ChildModel>, hcTT<ChildModel>, hc1AG<ChildModel>, NS<ChildModel>>> hc1_noart_aids_deaths;
-  TensorFixedSize <real_type, Sizes<hc2DS<ChildModel>, hcTT<ChildModel>, hc2AG<ChildModel>, NS<ChildModel>>> hc2_noart_aids_deaths;
-  TensorFixedSize <real_type, Sizes<hTS<ChildModel>, hc1DS<ChildModel>, hc1AG<ChildModel>, NS<ChildModel>>> hc1_art_aids_deaths;
-  TensorFixedSize <real_type, Sizes<hTS<ChildModel>, hc2DS<ChildModel>, hc2AG<ChildModel>, NS<ChildModel>>> hc2_art_aids_deaths;
-
-  real_type hc_art_num;
-
-  ChildModelState(const Parameters<ChildModel, real_type> &pars) {
-    reset();
-  }
-
-  void set_initial_state(const Parameters<ChildModel, real_type> &pars) {
-    reset();
-  }
-
-  void reset() {
-    hc1_hiv_pop.setZero();
-    hc2_hiv_pop.setZero();
-    hc1_art_pop.setZero();
-    hc2_art_pop.setZero();
-    hc1_noart_aids_deaths.setZero();
-    hc2_noart_aids_deaths.setZero();
-    hc1_art_aids_deaths.setZero();
-    hc2_art_aids_deaths.setZero();
-    hc_art_num = 0.0;
-  }
-};
-
-template<typename ModelVariant, typename real_type>
-struct State {
-  BaseModelState<ModelVariant, real_type> base;
-  ChildModelState<ModelVariant, real_type> children;
-
-  State(const Parameters<ModelVariant, real_type> &pars) :
-      base(pars),
-      children(pars) {}
-
-  void set_initial_state(const Parameters<ChildModel, real_type> &pars) {
-    base.set_initial_state();
-    children.set_initial_state();
-  }
-
-  void reset() {
-    base.reset();
-    children.reset();
-  }
-};
 
 namespace internal {
